@@ -1,14 +1,16 @@
 """PullRequest repository with domain-specific operations."""
 
 import uuid
-from datetime import datetime, timezone
-from typing import Any, Optional
+from datetime import UTC, datetime
+from typing import Any
 
-from sqlalchemy import and_, desc, func, or_, select
+from sqlalchemy import and_, desc, func, or_, select, text
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
+from sqlalchemy.orm.attributes import flag_modified
 
 from src.models import CheckRun, PRState, PullRequest, Repository, TriggerEvent
+
 from .base import BaseRepository
 
 
@@ -20,102 +22,89 @@ class PullRequestRepository(BaseRepository[PullRequest]):
         super().__init__(session, PullRequest)
 
     async def get_by_repo_and_number(
-        self, 
-        repository_id: uuid.UUID, 
-        pr_number: int
-    ) -> Optional[PullRequest]:
+        self, repository_id: uuid.UUID, pr_number: int
+    ) -> PullRequest | None:
         """Get PR by repository ID and PR number."""
         query = (
             select(PullRequest)
             .where(
                 and_(
                     PullRequest.repository_id == repository_id,
-                    PullRequest.pr_number == pr_number
+                    PullRequest.pr_number == pr_number,
                 )
             )
             .options(
                 selectinload(PullRequest.repository),
                 selectinload(PullRequest.check_runs),
-                selectinload(PullRequest.state_history)
+                selectinload(PullRequest.state_history),
             )
         )
         return await self._execute_single_query(query)
 
     async def get_by_repo_url_and_number(
-        self, 
-        repo_url: str, 
-        pr_number: int
-    ) -> Optional[PullRequest]:
+        self, repo_url: str, pr_number: int
+    ) -> PullRequest | None:
         """Get PR by repository URL and PR number."""
         query = (
             select(PullRequest)
             .join(Repository)
-            .where(
-                and_(
-                    Repository.url == repo_url,
-                    PullRequest.pr_number == pr_number
-                )
-            )
+            .where(and_(Repository.url == repo_url, PullRequest.pr_number == pr_number))
             .options(
                 selectinload(PullRequest.repository),
                 selectinload(PullRequest.check_runs),
-                selectinload(PullRequest.state_history)
+                selectinload(PullRequest.state_history),
             )
         )
         return await self._execute_single_query(query)
 
     async def get_active_prs_for_repo(
-        self, 
-        repository_id: uuid.UUID,
-        include_drafts: bool = False
+        self, repository_id: uuid.UUID, include_drafts: bool = False
     ) -> list[PullRequest]:
         """Get all active PRs for a repository."""
         conditions = [
             PullRequest.repository_id == repository_id,
-            PullRequest.state == PRState.OPENED
+            PullRequest.state == PRState.OPENED,
         ]
-        
+
         if not include_drafts:
-            conditions.append(PullRequest.draft == False)
-        
+            conditions.append(~PullRequest.draft)
+
         query = (
             select(PullRequest)
             .where(and_(*conditions))
             .order_by(desc(PullRequest.created_at))
             .options(
                 selectinload(PullRequest.repository),
-                selectinload(PullRequest.check_runs)
+                selectinload(PullRequest.check_runs),
             )
         )
         return await self._execute_query(query)
 
     async def get_prs_needing_check(
-        self, 
-        last_checked_before: datetime,
-        limit: Optional[int] = None
+        self, last_checked_before: datetime, limit: int | None = None
     ) -> list[PullRequest]:
         """Get PRs that need checking (haven't been checked recently)."""
         conditions = [
             PullRequest.state == PRState.OPENED,
             or_(
                 PullRequest.last_checked_at.is_(None),
-                PullRequest.last_checked_at < last_checked_before
-            )
+                PullRequest.last_checked_at < last_checked_before,
+            ),
         ]
-        
+
         query = (
             select(PullRequest)
             .where(and_(*conditions))
             .order_by(PullRequest.last_checked_at.asc().nulls_first())
             .options(
                 selectinload(PullRequest.repository),
-                selectinload(PullRequest.check_runs)
+                selectinload(PullRequest.check_runs),
             )
         )
-        
+
         if limit:
             query = query.limit(limit)
-            
+
         return await self._execute_query(query)
 
     async def update_state(
@@ -123,26 +112,30 @@ class PullRequestRepository(BaseRepository[PullRequest]):
         pr_id: uuid.UUID,
         new_state: PRState,
         trigger_event: TriggerEvent,
-        metadata: Optional[dict[str, Any]] = None
+        metadata: dict[str, Any] | None = None,
     ) -> PullRequest:
         """Update PR state and create state history record."""
         from .state_history import PRStateHistoryRepository
-        
+
         # Get the PR
         pr = await self.get_by_id_or_raise(pr_id)
         old_state = pr.state
-        
+
         # Validate transition
         if not pr.can_transition_to(new_state):
-            raise ValueError(f"Invalid state transition from {old_state} to {new_state}")
-        
+            raise ValueError(
+                f"Invalid state transition from {old_state} to {new_state}"
+            )
+
         # Update PR state
         pr.state = new_state
         if metadata:
             if pr.pr_metadata is None:
                 pr.pr_metadata = {}
             pr.pr_metadata.update(metadata)
-        
+            # Tell SQLAlchemy the JSONB field was modified
+            flag_modified(pr, "pr_metadata")
+
         # Create state history record
         history_repo = PRStateHistoryRepository(self.session)
         await history_repo.create_transition(
@@ -150,9 +143,9 @@ class PullRequestRepository(BaseRepository[PullRequest]):
             old_state=old_state,
             new_state=new_state,
             trigger_event=trigger_event,
-            metadata=metadata
+            metadata=metadata,
         )
-        
+
         await self.flush()
         await self.refresh(pr)
         return pr
@@ -160,124 +153,121 @@ class PullRequestRepository(BaseRepository[PullRequest]):
     async def mark_as_checked(self, pr_id: uuid.UUID) -> PullRequest:
         """Mark PR as checked (update last_checked_at)."""
         pr = await self.get_by_id_or_raise(pr_id)
-        pr.last_checked_at = datetime.now(timezone.utc)
+        pr.last_checked_at = datetime.now(UTC)
         await self.flush()
         await self.refresh(pr)
         return pr
 
     async def get_prs_with_failed_checks(
-        self, 
-        repository_id: Optional[uuid.UUID] = None,
-        limit: Optional[int] = None
+        self, repository_id: uuid.UUID | None = None, limit: int | None = None
     ) -> list[PullRequest]:
         """Get PRs that have failed check runs."""
         # Build subquery for PRs with failed checks
         failed_checks_subquery = (
             select(CheckRun.pr_id)
             .where(
-                and_(
-                    CheckRun.status == "completed",
-                    CheckRun.conclusion == "failure"
-                )
+                and_(CheckRun.status == "completed", CheckRun.conclusion == "failure")
             )
             .distinct()
         )
-        
+
         conditions = [
             PullRequest.state == PRState.OPENED,
-            PullRequest.id.in_(failed_checks_subquery)
+            PullRequest.id.in_(failed_checks_subquery),
         ]
-        
+
         if repository_id:
             conditions.append(PullRequest.repository_id == repository_id)
-        
+
         query = (
             select(PullRequest)
             .where(and_(*conditions))
             .order_by(desc(PullRequest.updated_at))
             .options(
                 selectinload(PullRequest.repository),
-                selectinload(PullRequest.check_runs)
+                selectinload(PullRequest.check_runs),
             )
         )
-        
+
         if limit:
             query = query.limit(limit)
-            
+
         return await self._execute_query(query)
 
     async def get_recent_prs(
         self,
         since: datetime,
-        repository_id: Optional[uuid.UUID] = None,
-        states: Optional[list[PRState]] = None,
-        limit: Optional[int] = None
+        repository_id: uuid.UUID | None = None,
+        states: list[PRState] | None = None,
+        limit: int | None = None,
     ) -> list[PullRequest]:
         """Get PRs created or updated since a given time."""
         conditions = [
-            or_(
-                PullRequest.created_at >= since,
-                PullRequest.updated_at >= since
-            )
+            or_(PullRequest.created_at >= since, PullRequest.updated_at >= since)
         ]
-        
+
         if repository_id:
             conditions.append(PullRequest.repository_id == repository_id)
-        
+
         if states:
             conditions.append(PullRequest.state.in_(states))
-        
+
         query = (
             select(PullRequest)
             .where(and_(*conditions))
             .order_by(desc(PullRequest.updated_at))
             .options(
                 selectinload(PullRequest.repository),
-                selectinload(PullRequest.check_runs)
+                selectinload(PullRequest.check_runs),
             )
         )
-        
+
         if limit:
             query = query.limit(limit)
-            
+
         return await self._execute_query(query)
 
     async def get_pr_statistics(
-        self, 
-        repository_id: Optional[uuid.UUID] = None
+        self, repository_id: uuid.UUID | None = None
     ) -> dict[str, Any]:
         """Get statistics about PRs."""
         # Base query conditions
         conditions = []
         if repository_id:
             conditions.append(PullRequest.repository_id == repository_id)
-        
+
         # Count by state
         state_counts = {}
         for state in PRState:
-            count_conditions = conditions + [PullRequest.state == state]
-            count_query = select(func.count(PullRequest.id)).where(and_(*count_conditions))
+            count_conditions = [*conditions, PullRequest.state == state]
+            count_query = select(func.count(PullRequest.id)).where(
+                and_(*count_conditions)
+            )
             result = await self.session.execute(count_query)
             state_counts[state.value] = result.scalar_one()
-        
+
         # Count active (opened) PRs
-        active_conditions = conditions + [
+        active_conditions = [
+            *conditions,
             PullRequest.state == PRState.OPENED,
-            PullRequest.draft == False
+            ~PullRequest.draft,
         ]
-        active_query = select(func.count(PullRequest.id)).where(and_(*active_conditions))
+        active_query = select(func.count(PullRequest.id)).where(
+            and_(*active_conditions)
+        )
         result = await self.session.execute(active_query)
         active_count = result.scalar_one()
-        
+
         # Count draft PRs
-        draft_conditions = conditions + [
+        draft_conditions = [
+            *conditions,
             PullRequest.state == PRState.OPENED,
-            PullRequest.draft == True
+            PullRequest.draft,
         ]
         draft_query = select(func.count(PullRequest.id)).where(and_(*draft_conditions))
         result = await self.session.execute(draft_query)
         draft_count = result.scalar_one()
-        
+
         return {
             "total": sum(state_counts.values()),
             "by_state": state_counts,
@@ -287,71 +277,69 @@ class PullRequestRepository(BaseRepository[PullRequest]):
 
     async def search_prs(
         self,
-        query_text: Optional[str] = None,
-        author: Optional[str] = None,
-        state: Optional[PRState] = None,
-        repository_id: Optional[uuid.UUID] = None,
-        limit: Optional[int] = None,
-        offset: Optional[int] = None
+        query_text: str | None = None,
+        author: str | None = None,
+        state: PRState | None = None,
+        repository_id: uuid.UUID | None = None,
+        limit: int | None = None,
+        offset: int | None = None,
     ) -> list[PullRequest]:
         """Search PRs with various filters."""
         conditions = []
-        
+
         if query_text:
             conditions.append(
                 or_(
                     PullRequest.title.ilike(f"%{query_text}%"),
-                    PullRequest.body.ilike(f"%{query_text}%")
+                    PullRequest.body.ilike(f"%{query_text}%"),
                 )
             )
-        
+
         if author:
             conditions.append(PullRequest.author.ilike(f"%{author}%"))
-        
+
         if state:
             conditions.append(PullRequest.state == state)
-        
+
         if repository_id:
             conditions.append(PullRequest.repository_id == repository_id)
-        
+
         query = (
             select(PullRequest)
-            .where(and_(*conditions) if conditions else True)
+            .where(and_(*conditions) if conditions else text("1=1"))
             .order_by(desc(PullRequest.updated_at))
             .options(
                 selectinload(PullRequest.repository),
-                selectinload(PullRequest.check_runs)
+                selectinload(PullRequest.check_runs),
             )
         )
-        
+
         if offset:
             query = query.offset(offset)
-        
+
         if limit:
             query = query.limit(limit)
-            
+
         return await self._execute_query(query)
 
     async def bulk_update_last_checked(
-        self, 
-        pr_ids: list[uuid.UUID], 
-        checked_at: Optional[datetime] = None
+        self, pr_ids: list[uuid.UUID], checked_at: datetime | None = None
     ) -> int:
         """Bulk update last_checked_at for multiple PRs."""
         if not pr_ids:
             return 0
-            
+
         if checked_at is None:
-            checked_at = datetime.now(timezone.utc)
-        
+            checked_at = datetime.now(UTC)
+
         from sqlalchemy import update
-        
+
         stmt = (
             update(PullRequest)
             .where(PullRequest.id.in_(pr_ids))
             .values(last_checked_at=checked_at)
         )
-        
+
         result = await self.session.execute(stmt)
         await self.flush()
         return result.rowcount
