@@ -2,8 +2,9 @@
 
 import asyncio
 import logging
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Any, AsyncGenerator, Optional
+from typing import Any
 
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -13,6 +14,7 @@ logger = logging.getLogger(__name__)
 
 class TransactionError(Exception):
     """Exception raised during transaction operations."""
+
     pass
 
 
@@ -21,7 +23,7 @@ class DatabaseTransaction:
 
     def __init__(self, session: AsyncSession, auto_commit: bool = True):
         """Initialize transaction manager.
-        
+
         Args:
             session: Database session to manage
             auto_commit: Whether to auto-commit on successful completion
@@ -42,15 +44,17 @@ class DatabaseTransaction:
             logger.error(f"Failed to begin transaction: {e}")
             raise TransactionError(f"Failed to begin transaction: {e}") from e
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb) -> None:
+    async def __aexit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
         """Exit transaction context with automatic rollback on errors."""
         try:
             if exc_type is not None:
                 # Exception occurred, rollback
                 await self.rollback()
-                logger.warning(f"Transaction rolled back due to {exc_type.__name__}: {exc_val}")
-                return False  # Re-raise the exception
-            
+                logger.warning(
+                    f"Transaction rolled back due to {exc_type.__name__}: {exc_val}"
+                )
+                return None  # Re-raise the exception
+
             # No exception, commit if auto_commit is enabled
             if self.auto_commit and not self._committed:
                 await self.commit()
@@ -63,7 +67,7 @@ class DatabaseTransaction:
         """Manually commit the transaction."""
         if self._rolled_back:
             raise TransactionError("Cannot commit after rollback")
-        
+
         if not self._committed:
             try:
                 await self.session.commit()
@@ -89,7 +93,7 @@ class DatabaseTransaction:
         """Flush pending changes without committing."""
         if self._rolled_back:
             raise TransactionError("Cannot flush after rollback")
-        
+
         try:
             await self.session.flush()
         except Exception as e:
@@ -99,18 +103,17 @@ class DatabaseTransaction:
 
 @asynccontextmanager
 async def database_transaction(
-    session: AsyncSession, 
-    auto_commit: bool = True
+    session: AsyncSession, auto_commit: bool = True
 ) -> AsyncGenerator[AsyncSession, None]:
     """Context manager for database transactions.
-    
+
     Args:
         session: Database session to use
         auto_commit: Whether to automatically commit on success
-        
+
     Yields:
         The database session within transaction context
-        
+
     Example:
         async with database_transaction(session) as tx_session:
             # Perform database operations
@@ -132,10 +135,10 @@ class RetryableTransaction:
         max_retries: int = 3,
         base_delay: float = 0.1,
         backoff_factor: float = 2.0,
-        auto_commit: bool = True
+        auto_commit: bool = True,
     ):
         """Initialize retryable transaction.
-        
+
         Args:
             session: Database session to use
             max_retries: Maximum number of retry attempts
@@ -149,40 +152,45 @@ class RetryableTransaction:
         self.backoff_factor = backoff_factor
         self.auto_commit = auto_commit
 
-    async def execute(self, operation) -> Any:
+    async def execute(self, operation: Any) -> Any:
         """Execute operation with retry logic.
-        
+
         Args:
             operation: Async callable that performs database operations
-            
+
         Returns:
             Result of the operation
-            
+
         Raises:
             TransactionError: After all retries are exhausted
         """
         last_exception = None
-        
+
         for attempt in range(self.max_retries + 1):
             try:
-                async with database_transaction(self.session, self.auto_commit) as session:
+                async with database_transaction(
+                    self.session, self.auto_commit
+                ) as session:
                     result = await operation(session)
                     return result
             except SQLAlchemyError as e:
                 last_exception = e
                 if attempt < self.max_retries:
-                    delay = self.base_delay * (self.backoff_factor ** attempt)
+                    delay = self.base_delay * (self.backoff_factor**attempt)
                     logger.warning(
-                        f"Transaction attempt {attempt + 1} failed, retrying in {delay}s: {e}"
+                        f"Transaction attempt {attempt + 1} failed, "
+                        f"retrying in {delay}s: {e}"
                     )
                     await asyncio.sleep(delay)
                 else:
-                    logger.error(f"All {self.max_retries + 1} transaction attempts failed")
+                    logger.error(
+                        f"All {self.max_retries + 1} transaction attempts failed"
+                    )
             except Exception as e:
                 # Non-retryable exception
                 logger.error(f"Non-retryable error in transaction: {e}")
                 raise TransactionError(f"Non-retryable transaction error: {e}") from e
-        
+
         # All retries exhausted
         raise TransactionError(
             f"Transaction failed after {self.max_retries + 1} attempts"
@@ -195,20 +203,20 @@ async def retryable_transaction(
     max_retries: int = 3,
     base_delay: float = 0.1,
     backoff_factor: float = 2.0,
-    auto_commit: bool = True
+    auto_commit: bool = True,
 ) -> AsyncGenerator[RetryableTransaction, None]:
     """Context manager for retryable transactions.
-    
+
     Args:
         session: Database session to use
         max_retries: Maximum number of retry attempts
         base_delay: Base delay between retries in seconds
         backoff_factor: Exponential backoff multiplier
         auto_commit: Whether to auto-commit on success
-        
+
     Yields:
         RetryableTransaction instance
-        
+
     Example:
         async with retryable_transaction(session) as tx:
             result = await tx.execute(
@@ -221,13 +229,13 @@ async def retryable_transaction(
     yield tx
 
 
-def transactional(auto_commit: bool = True, retries: int = 0):
+def transactional(auto_commit: bool = True, retries: int = 0) -> Any:
     """Decorator for automatic transaction management.
-    
+
     Args:
         auto_commit: Whether to auto-commit on success
         retries: Number of retries for transient failures
-        
+
     Example:
         @transactional(retries=3)
         async def complex_operation(self, session: AsyncSession, ...):
@@ -235,31 +243,33 @@ def transactional(auto_commit: bool = True, retries: int = 0):
             await self.repo1.create(...)
             await self.repo2.update(...)
     """
-    def decorator(func):
-        async def wrapper(*args, **kwargs):
+
+    def decorator(func: Any) -> Any:
+        async def wrapper(*args: Any, **kwargs: Any) -> Any:
             # Extract session from args or kwargs
             session = None
             for arg in args:
                 if isinstance(arg, AsyncSession):
                     session = arg
                     break
-            
+
             if session is None:
-                session = kwargs.get('session')
-            
+                session = kwargs.get("session")
+
             if session is None:
                 raise ValueError("No AsyncSession found in arguments")
-            
+
             if retries > 0:
                 async with retryable_transaction(
                     session, max_retries=retries, auto_commit=auto_commit
                 ) as tx:
                     return await tx.execute(lambda s: func(*args, **kwargs))
             else:
-                async with database_transaction(session, auto_commit) as tx_session:
+                async with database_transaction(session, auto_commit):
                     return await func(*args, **kwargs)
-        
+
         return wrapper
+
     return decorator
 
 
@@ -268,7 +278,7 @@ class UnitOfWork:
 
     def __init__(self, session: AsyncSession):
         """Initialize unit of work.
-        
+
         Args:
             session: Database session to use
         """
@@ -276,9 +286,9 @@ class UnitOfWork:
         self._operations: list[tuple[str, Any, dict]] = []
         self._committed = False
 
-    def add_operation(self, operation_type: str, entity: Any, **kwargs) -> None:
+    def add_operation(self, operation_type: str, entity: Any, **kwargs: Any) -> None:
         """Add an operation to the unit of work.
-        
+
         Args:
             operation_type: Type of operation ('create', 'update', 'delete')
             entity: Entity to operate on
@@ -286,28 +296,28 @@ class UnitOfWork:
         """
         if self._committed:
             raise TransactionError("Cannot add operations after commit")
-        
+
         self._operations.append((operation_type, entity, kwargs))
 
     async def commit(self) -> None:
         """Execute all operations in a single transaction."""
         if self._committed:
             raise TransactionError("Unit of work already committed")
-        
+
         async with database_transaction(self.session) as session:
-            for operation_type, entity, kwargs in self._operations:
-                if operation_type == 'create':
+            for operation_type, entity, _kwargs in self._operations:
+                if operation_type == "create":
                     session.add(entity)
-                elif operation_type == 'update':
-                    session.merge(entity)
-                elif operation_type == 'delete':
+                elif operation_type == "update":
+                    await session.merge(entity)
+                elif operation_type == "delete":
                     await session.delete(entity)
                 else:
                     raise ValueError(f"Unknown operation type: {operation_type}")
-            
+
             # Flush to ensure all operations are applied
             await session.flush()
-        
+
         self._committed = True
         logger.debug(f"Unit of work committed {len(self._operations)} operations")
 
@@ -320,13 +330,13 @@ class UnitOfWork:
 @asynccontextmanager
 async def unit_of_work(session: AsyncSession) -> AsyncGenerator[UnitOfWork, None]:
     """Context manager for unit of work pattern.
-    
+
     Args:
         session: Database session to use
-        
+
     Yields:
         UnitOfWork instance
-        
+
     Example:
         async with unit_of_work(session) as uow:
             uow.add_operation('create', new_entity)
